+++ tests/Provider/QuoterProvider.php
@@ -91,18 +91,17 @@ class QuoterProvider
         ];
     }
 
-    /**
-     * @return string[][]
-     */
     public static function tableNameParts(): array
     {
         return [
-            ['animal', 'animal',],
-            ['dbo.animal', 'animal', 'dbo'],
-            ['[dbo].[animal]', 'animal', 'dbo'],
-            ['[other].[animal2]', 'animal2', 'other'],
-            ['other.[animal2]', 'animal2', 'other'],
-            ['other.animal2', 'animal2', 'other'],
+            ['', ['name' => '']],
+            ['[]', ['name' => '']],
+            ['animal', ['name' => 'animal']],
+            ['[animal]', ['name' => 'animal']],
+            ['dbo.animal', ['schemaName' => 'dbo', 'name' => 'animal']],
+            ['[dbo].[animal]', ['schemaName' => 'dbo', 'name' => 'animal']],
+            ['[dbo].animal', ['schemaName' => 'dbo', 'name' => 'animal']],
+            ['dbo.[animal]', ['schemaName' => 'dbo', 'name' => 'animal']],
         ];
     }